用户实体类更新

+++ src/main/java/com/mpic/evolution/chair/pojo/entity/EcmUser.java
@@ -80,7 +80,7 @@ public class EcmUser implements Serializable {
     /**
      * 用户头像状态 0 未审核，1 通过 ，2 不通过
      */
-    private Byte userLogoStatus;
+    private Short userLogoStatus;
 
     /**
      * 用户生日
